Split Single Topic Page into smaller Components (#34)
