+++ ui/src/app/layout/topics/deprecation/deprecation.component.html
@@ -1,0 +1,36 @@
+<div
+    *ngIf="topic?.topicType !== 'INTERNAL' && isOwnerOfTopic  && !(selectedEnvironment | async)?.stagingOnly"
+    class="card mb-3">
+    <div class="card-header bg-hermes-deprecation"><i
+        class="fas fa-ghost mr-3"></i>{{ 'Deprecation' | translate }}</div>
+    <div class="card-body">
+        <p id="deprecationHtml" [innerHTML]="deprecateTopicHtml | async"></p>
+        <ngb-alert type="danger" [dismissible]="false"><span
+            [innerHTML]="'DEPRECATE_ALL_STAGES_ALERT_HTML' | translate"></span>
+        </ngb-alert>
+        <div class="form-group">
+            <label>{{ 'Deprecation Description' | translate }}</label>
+            <textarea class="form-control" rows="2"
+                      [(ngModel)]="deprecatedDescription"></textarea><br>
+            <label>{{ 'PICK_DEPRECATE_DATE' | translate }}</label><br>
+            <form class="form-inline">
+                <div class="form-group">
+                    <div class="input-group">
+                        <input class="form-control" [(ngModel)]="eolDate" ngbDatepicker
+                               [minDate]="minDeprecationDate | async"
+                               name="eolDate" #p="ngbDatepicker">
+                        <div class="input-group-append">
+                            <button class="fas fa-calendar-alt" (click)="p.toggle()"
+                                    type="button"></button>
+                        </div>
+                    </div>
+                </div>
+            </form>
+        </div>
+        <button [disabled]="!(eolDate && deprecatedDescription)" appSpinnerWhile
+                #d="appSpinnerWhile"
+                class="btn btn-warning"
+                (click)="d.while(handleDeprecationRequest())"><i
+            class="fas fa-ghost mr-3"></i>{{ 'Deprecate Topic Label' | translate }}</button>
+    </div>
+</div>
