+++ ui/src/app/layout/topics/deprecation/deprecation.component.scss
@@ -1,0 +1,4 @@
+.bg-hermes-deprecation {
+    color: #000000;
+    background-color: #c0c0c0;
+}
