+++ ui/src/app/layout/topics/schemasection/schema-section.component.html
@@ -15,7 +15,7 @@
                         <button id="schemaDeleteButton" appSpinnerWhile
                                 #p="appSpinnerWhile" (click)="(openDeleteConfirmDlg(deleteLatestSchemaDlg))"
                                 class="btn btn-danger ml-3"
-                                *ngIf="selectedSchemaVersion.isLatest && isOwnerOfTopic && (topicSubscribers.length === 0
+                                *ngIf="selectedSchemaVersion.isLatest && isOwnerOfTopic && (topicSubscribers?.length === 0
                                 || (!(selectedEnvironment | async)?.stagingOnly && (schemaDeleteWithSub | async)))">
                             <i class="fas fa-trash mr-3"></i>{{ 'Delete Schema' | translate }}
                         </button>
