+++ ui/src/app/layout/topics/schemasection/schema-section.component.scss
@@ -1,0 +1,5 @@
+.scrollbar {
+    overflow: auto;
+    width: 100%;
+    max-height: 300px;
+}
