+++ ui/src/app/layout/topics/topicmetadatatable/topic-metadata-table.component.scss
@@ -1,0 +1,10 @@
+.deprecatedNote {
+    font-weight: bold;
+    font-style: italic;
+    margin-left: 10px;
+}
+
+.deprecationText {
+    font-style: italic;
+    margin-left: 10px;
+}
