+++ ui/src/app/layout/topics/topics.module.ts
@@ -11,7 +11,11 @@ import { SingleTopicComponent } from './single-topic.component';
 import { SpinnerWhileModule } from '../../shared/modules/spinner-while/spinner-while.module';
 import { HighlightModule } from 'ngx-highlightjs';
 import { TopicConfigEditorComponent } from './topic-config-editor.component';
-import { SchemaSectionComponent } from './schema-section.component';
+import { SchemaSectionComponent } from './schemasection/schema-section.component';
+import { TopicMetadataTableComponent } from './topicmetadatatable/topic-metadata-table.component';
+import { SubscriptionSectionComponent } from './subscribesection/subscribe-section.component';
+import { DeprecationComponent } from './deprecation/deprecation.component';
+import { DeleteTopicComponent } from './deletetopic/delete-topic.component';
 
 export const getHighlightLanguages = () => ({
     json: () => import('highlight.js/lib/languages/json')
@@ -20,7 +24,9 @@ export const getHighlightLanguages = () => ({
 @NgModule({
     imports: [CommonModule, TopicsRoutingModule, TranslateModule, FormsModule, NgbModule,
         SpinnerWhileModule, HighlightModule],
-    declarations: [TopicsComponent, TableSortDirective, SingleTopicComponent, TopicConfigEditorComponent, SchemaSectionComponent]
+    declarations: [TopicsComponent, TableSortDirective, SingleTopicComponent,
+        TopicConfigEditorComponent, SchemaSectionComponent, TopicMetadataTableComponent,
+        SubscriptionSectionComponent, DeprecationComponent, DeleteTopicComponent]
 })
 export class TopicsModule {
 }
